object: invalidate signal connection later
This defers the call of g_closure_invalidate() until the end of a garbage
collection. We can't stop tracing signal connections in the middle of
garbage collection, and we can't defer the actual freeing of the
closures because that caused too many race conditions with the garbage
collector. Instead, defer the invalidation.

This probably means that more JS objects will survive a garbage
collection when they didn't have to, but they should get collected in the
following garbage collection.